git-svn-id: https://gorgonlib.googlecode.com/svn/branches/2.x@730 a61f9e97-bd4c-0410-8c65-85b29eaa70ff

+++ Dorian/Gorgon/Gorgon.Graphics/Targets/GorgonSwapChain.cs
@@ -373,7 +373,7 @@ namespace GorgonLibrary.Graphics
 				}
 				else
 				{
-					if (sdEx.ResultCode.Code == (int)GI.DXGIError.NotCurrentlyAvailable)
+					if (sdEx.ResultCode.Code == GI.DXGIError.NotCurrentlyAvailable)
 						Gorgon.Log.Print("GorgonSwapChain '{0}': Could not switch to full screen mode because the device is not currently available.  Possible causes are:  .", GorgonLoggingLevel.All, Name);
 					else
 						throw sdEx;
